Fix sendContent, merge fixes
