+++ LeanplumSDK/LeanplumSDK/ClassesSwift/Actions/ActionManager+Definition.swift
@@ -13,6 +13,10 @@ extension ActionManager {
         definitions.append(definition)
     }
     
+    @objc public func definition(withName name:String) -> ActionDefinition? {
+        return self.definitions.first(where: { $0.name == name })
+    }
+    
     func getActionDefinitionType(name: String) -> UInt {
         let definition = definition(withName: name)
         if let definition = definition {
