+++ LeanplumSDK/LeanplumSDK/ClassesSwift/Actions/ActionManager.swift
@@ -10,17 +10,14 @@ import Foundation
 @objc public class ActionManager: NSObject {
     @objc public static let shared: ActionManager = .init()
     
-    // TODO: add contentVersion
-
     lazy var queue: Queue = Queue()
     lazy var scheduler: Scheduler = Scheduler()
     lazy var state = State()
     
-    @objc public lazy var definitions: [ActionDefinition] = []
-    // Move to struct
-    @objc public lazy var messages: [AnyHashable:Any] = [:]
-    @objc public lazy var messagesDataFromServer: [AnyHashable:Any] = [:] // messageDiffs
-    @objc public lazy var devModeActionDefinitionsFromServer: [AnyHashable:Any] = [:]
+    @objc public var definitions: [ActionDefinition] = []
+    @objc public var messages: [AnyHashable:Any] = [:]
+    @objc public var messagesDataFromServer: [AnyHashable:Any] = [:] // messageDiffs
+    @objc public var devModeActionDefinitionsFromServer: [AnyHashable:Any] = [:]
 
     public var enabled: Bool = true
